+++ tests/loops/fixpoint_convergence.m
@@ -3,7 +3,7 @@
 % With --fixpoint: join of scalar and matrix[2 x 1], matrix[3 x 1], etc. converges to matrix[None x 1]
 % EXPECT: warnings = 1
 % EXPECT: A = matrix[2 x 1]
-% EXPECT_FIXPOINT: warnings = 3
+% EXPECT_FIXPOINT: warnings = 1
 % EXPECT_FIXPOINT: A = unknown
 
 A = 1;